se agregan pruebas unitarias

+++ src/main/java/com/evaluaciones/dao/EstudianteDao.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.evaluaciones.model.EstudianteModel;
 
-public interface EstudianteDao extends CrudRepository<EstudianteModel, Integer> {
+public interface EstudianteDao extends CrudRepository<EstudianteModel, Long> {
 
 	List<EstudianteModel> findByNombreAndApellidoPaternoAndApellidoMaterno 
 		(String nombre, String apellidoPaterno, String apellidoMaterno);
